+++ keyboards/lily58/glow_enc/info.json
@@ -8,13 +8,16 @@
         "pid": "0x0059",
         "device_version": "1.0.0"
     },
+    "features": {
+      "rgblight": true
+    },
     "encoder": {
+        "enabled": true,
         "rotary": [
             {"pin_a": "F5", "pin_b": "D4"}
         ]
     },
     "split": {
-        "soft_serial_pin": "D2",
         "encoder": {
             "right": {
                 "rotary": [
@@ -23,17 +26,18 @@
             }
         }
     },
-    "processor": "atmega32u4",
-    "bootloader": "caterina",
-    "layouts": {
-        "LAYOUT": {
-            "layout": [
-                {"x":0, "y":0.5}, {"x":1, "y":0.375}, {"x":2, "y":0.125}, {"x":3, "y":0}, {"x":4, "y":0.125}, {"x":5, "y":0.25}, {"x":10.5, "y":0.25}, {"x":11.5, "y":0.125}, {"x":12.5, "y":0}, {"x":13.5, "y":0.125}, {"x":14.5, "y":0.375}, {"x":15.5, "y":0.5},
-                {"x":0, "y":1.5}, {"x":1, "y":1.375}, {"x":2, "y":1.125}, {"x":3, "y":1}, {"x":4, "y":1.125}, {"x":5, "y":1.25}, {"x":10.5, "y":1.25}, {"x":11.5, "y":1.125}, {"x":12.5, "y":1}, {"x":13.5, "y":1.125}, {"x":14.5, "y":1.375}, {"x":15.5, "y":1.5},
-                {"x":0, "y":2.5}, {"x":1, "y":2.375}, {"x":2, "y":2.125}, {"x":3, "y":2}, {"x":4, "y":2.125}, {"x":5, "y":2.25}, {"x":10.5, "y":2.25}, {"x":11.5, "y":2.125}, {"x":12.5, "y":2}, {"x":13.5, "y":2.125}, {"x":14.5, "y":2.375}, {"x":15.5, "y":2.5},
-                {"x":0, "y":3.5}, {"x":1, "y":3.375}, {"x":2, "y":3.125}, {"x":3, "y":3}, {"x":4, "y":3.125}, {"x":5, "y":3.25}, {"x":6, "y":2.75}, {"x":9.5, "y":2.75}, {"x":10.5, "y":3.25}, {"x":11.5, "y":3.125}, {"x":12.5, "y":3}, {"x":13.5, "y":3.125}, {"x":14.5, "y":3.375}, {"x":15.5, "y":3.5},
-                {"x":2.5, "y":4.125}, {"x":3.5, "y":4.15}, {"x":4.5, "y":4.25}, {"x":6, "y":4.25, "h":1.5}, {"x":9.5, "y":4.25, "h":1.5}, {"x":11, "y":4.25}, {"x":12, "y":4.15}, {"x":13, "y":4.15}
-            ]
+    "rgblight": {
+        "led_count": 72,
+        "pin": "F4",
+        "split": true,
+        "split_count": [36, 36],
+        "max_brightness": 120,
+        "animations": {
+            "rainbow_swirl": true,
+            "static_gradient": true,
+            "twinkle": true
         }
-    }
+    },
+    "processor": "atmega32u4",
+    "bootloader": "caterina"
 }
