+++ keyboards/lily58/glow_enc/rules.mk
@@ -1,8 +1,1 @@
-ENCODER_ENABLE = yes       # ENables the use of one or more encoders
-RGBLIGHT_ENABLE = yes      # Enable keyboard RGB light
-LTO_ENABLE = yes 		   # significantly reduce the compiled size, but disable the legacy TMK Macros and Functions features
-BOOTMAGIC_ENABLE = no
-RGB_MATRIX_ENABLE = no     # Per key rgb leds
-
-# Do not enable SLEEP_LED_ENABLE. it uses the same timer as BACKLIGHT_ENABLE
-SLEEP_LED_ENABLE = no    # Breathing sleep LED during USB suspend
+# This file intentionally left blank
