+++ keyboards/lily58/info.json
@@ -1,0 +1,83 @@
+{
+  "feautres": {
+    "bootmagic": true,
+    "mousekey": true,
+    "extrakey": true
+  },
+  "build": {
+    "lto": true
+  },
+  "matrix_pins": {
+    "cols": ["F6", "F7", "B1", "B3", "B2", "B6"],
+    "rows": ["C6", "D7", "E6", "B4", "B5"]
+  },
+  "split": {
+    "enabled": true,
+    "soft_serial_pin": "D2"
+  },
+  "diode_direction": "COL2ROW",
+  "layouts": {
+    "LAYOUT": {
+      "layout": [
+        { "matrix": [0, 0], "w": 1, "x": 0, "y": 0.5 },
+        { "matrix": [0, 1], "w": 1, "x": 1, "y": 0.375 },
+        { "matrix": [0, 2], "w": 1, "x": 2, "y": 0.125 },
+        { "matrix": [0, 3], "w": 1, "x": 3, "y": 0 },
+        { "matrix": [0, 4], "w": 1, "x": 4, "y": 0.125 },
+        { "matrix": [0, 5], "w": 1, "x": 5, "y": 0.25 },
+        { "matrix": [5, 0], "w": 1, "x": 10.5, "y": 0.25 },
+        { "matrix": [5, 1], "w": 1, "x": 11.5, "y": 0.125 },
+        { "matrix": [5, 2], "w": 1, "x": 12.5, "y": 0 },
+        { "matrix": [5, 3], "w": 1, "x": 13.5, "y": 0.125 },
+        { "matrix": [5, 4], "w": 1, "x": 14.5, "y": 0.375 },
+        { "matrix": [5, 5], "w": 1, "x": 15.5, "y": 0.5 },
+        { "matrix": [1, 0], "w": 1, "x": 0, "y": 1.5 },
+        { "matrix": [1, 1], "w": 1, "x": 1, "y": 1.375 },
+        { "matrix": [1, 2], "w": 1, "x": 2, "y": 1.125 },
+        { "matrix": [1, 3], "w": 1, "x": 3, "y": 1 },
+        { "matrix": [1, 4], "w": 1, "x": 4, "y": 1.125 },
+        { "matrix": [1, 5], "w": 1, "x": 5, "y": 1.25 },
+        { "matrix": [6, 0], "w": 1, "x": 10.5, "y": 1.25 },
+        { "matrix": [6, 1], "w": 1, "x": 11.5, "y": 1.125 },
+        { "matrix": [6, 2], "w": 1, "x": 12.5, "y": 1 },
+        { "matrix": [6, 3], "w": 1, "x": 13.5, "y": 1.125 },
+        { "matrix": [6, 4], "w": 1, "x": 14.5, "y": 1.375 },
+        { "matrix": [6, 5], "w": 1, "x": 15.5, "y": 1.5 },
+        { "matrix": [2, 0], "w": 1, "x": 0, "y": 2.5 },
+        { "matrix": [2, 1], "w": 1, "x": 1, "y": 2.375 },
+        { "matrix": [2, 2], "w": 1, "x": 2, "y": 2.125 },
+        { "matrix": [2, 3], "w": 1, "x": 3, "y": 2 },
+        { "matrix": [2, 4], "w": 1, "x": 4, "y": 2.125 },
+        { "matrix": [2, 5], "w": 1, "x": 5, "y": 2.25 },
+        { "matrix": [7, 0], "w": 1, "x": 10.5, "y": 2.25 },
+        { "matrix": [7, 1], "w": 1, "x": 11.5, "y": 2.125 },
+        { "matrix": [7, 2], "w": 1, "x": 12.5, "y": 2 },
+        { "matrix": [7, 3], "w": 1, "x": 13.5, "y": 2.125 },
+        { "matrix": [7, 4], "w": 1, "x": 14.5, "y": 2.375 },
+        { "matrix": [7, 5], "w": 1, "x": 15.5, "y": 2.5 },
+        { "matrix": [3, 0], "w": 1, "x": 0, "y": 3.5 },
+        { "matrix": [3, 1], "w": 1, "x": 1, "y": 3.375 },
+        { "matrix": [3, 2], "w": 1, "x": 2, "y": 3.125 },
+        { "matrix": [3, 3], "w": 1, "x": 3, "y": 3 },
+        { "matrix": [3, 4], "w": 1, "x": 4, "y": 3.125 },
+        { "matrix": [3, 5], "w": 1, "x": 5, "y": 3.25 },
+        { "matrix": [4, 5], "w": 1, "x": 6, "y": 2.75 },
+        { "matrix": [8, 0], "w": 1, "x": 9.5, "y": 2.75 },
+        { "matrix": [8, 1], "w": 1, "x": 10.5, "y": 3.25 },
+        { "matrix": [8, 2], "w": 1, "x": 11.5, "y": 3.125 },
+        { "matrix": [8, 3], "w": 1, "x": 12.5, "y": 3 },
+        { "matrix": [8, 4], "w": 1, "x": 13.5, "y": 3.125 },
+        { "matrix": [8, 5], "w": 1, "x": 14.5, "y": 3.375 },
+        { "matrix": [9, 5], "w": 1, "x": 15.5, "y": 3.5 },
+        { "matrix": [4, 1], "w": 1, "x": 2.5, "y": 4.125 },
+        { "matrix": [4, 2], "w": 1, "x": 3.5, "y": 4.15 },
+        { "matrix": [4, 3], "w": 1, "x": 4.5, "y": 4.25 },
+        { "matrix": [4, 4], "w": 1, "x": 6, "y": 4.25 },
+        { "matrix": [9, 1], "w": 1, "x": 9.5, "y": 4.25 },
+        { "matrix": [9, 2], "w": 1, "x": 11, "y": 4.25 },
+        { "matrix": [9, 3], "w": 1, "x": 12, "y": 4.15 },
+        { "matrix": [9, 4], "w": 1, "x": 13, "y": 4.15 }
+      ]
+    }
+  }
+}
