+++ keyboards/lily58/keymaps/lily58l/keymap.c
@@ -18,18 +18,12 @@
 
 #include QMK_KEYBOARD_H
 
-
-extern uint8_t is_master;
-
 enum layers {
     _QWERTY,
     _LOWER,
     _RAISE,
-    _ADJUST,
+    _ADJUST
 };
-
-#define RAISE MO(_RAISE)
-#define LOWER MO(_LOWER)
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 
@@ -53,7 +47,7 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
   KC_TAB,   KC_Q,   KC_W,    KC_E,    KC_R,    KC_T,                     KC_Y,    KC_U,    KC_I,    KC_O,    KC_P,    KC_MINS,
   KC_LCTL,  KC_A,   KC_S,    KC_D,    KC_F,    KC_G,                     KC_H,    KC_J,    KC_K,    KC_L,    KC_SCLN, KC_QUOT,
   KC_LSFT,  KC_Z,   KC_X,    KC_C,    KC_V,    KC_B, KC_LBRC,  KC_RBRC,  KC_N,    KC_M,    KC_COMM, KC_DOT,  KC_SLSH,  KC_RSFT,
-                             KC_LALT, KC_LGUI,LOWER, KC_SPC,   KC_ENT,   RAISE,   KC_BSPC, KC_RGUI
+                             KC_LALT, KC_LGUI,TL_LOWR, KC_SPC,   KC_ENT,   TL_UPPR,   KC_BSPC, KC_RGUI
 ),
 /* LOWER
  * ,-----------------------------------------.                    ,-----------------------------------------.
@@ -121,12 +115,6 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
   )
 };
 
-layer_state_t layer_state_set_user(layer_state_t state) {
-    state = update_tri_layer_state(state, _RAISE, _LOWER, _ADJUST);
-    return state;
-}
-
-//SSD1306 OLED update loop, make sure to enable OLED_ENABLE=yes in rules.mk
 #ifdef OLED_ENABLE
 
 oled_rotation_t oled_init_user(oled_rotation_t rotation) {
@@ -174,47 +162,6 @@ void render_lily58_logo(void) {
     0x00, 0x00, 0x00, 0x00, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x00
 };
     oled_write_raw_P(lily58_logo, sizeof(lily58_logo));
-}
-
-
-#    define KEYLOG_LEN 6
-char     keylog_str[KEYLOG_LEN] = {};
-uint8_t  keylogs_str_idx        = 0;
-uint16_t log_timer              = 0;
-
-const char code_to_name[60] = {
-    ' ', ' ', ' ', ' ', 'a', 'b', 'c', 'd', 'e', 'f',
-    'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p',
-    'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z',
-    '1', '2', '3', '4', '5', '6', '7', '8', '9', '0',
-    'R', 'E', 'B', 'T', '_', '-', '=', '[', ']', '\\',
-    '#', ';', '\'', '`', ',', '.', '/', ' ', ' ', ' '};
-
-void add_keylog(uint16_t keycode) {
-    if ((keycode >= QK_MOD_TAP && keycode <= QK_MOD_TAP_MAX) || (keycode >= QK_LAYER_TAP && keycode <= QK_LAYER_TAP_MAX)) {
-        keycode = keycode & 0xFF;
-    }
-
-    for (uint8_t i = KEYLOG_LEN - 1; i > 0; i--) {
-        keylog_str[i] = keylog_str[i - 1];
-    }
-    if (keycode < 60) {
-        keylog_str[0] = code_to_name[keycode];
-    }
-    keylog_str[KEYLOG_LEN - 1] = 0;
-
-    log_timer = timer_read();
-}
-
-void update_log(void) {
-    if (timer_elapsed(log_timer) > 750) {
-        add_keylog(0);
-    }
-}
-
-void render_keylogger_status(void) {
-    oled_write_P(PSTR("KLogr"), false);
-    oled_write(keylog_str, false);
 }
 
 void render_default_layer_state(void) {
@@ -268,11 +215,11 @@ void render_status_main(void) {
     render_mod_status(get_mods());
     // Add a empty line
     oled_write_P(PSTR("-----"), false);
-    render_keylogger_status();
+    oled_write_P(PSTR("KLogr"), false);
+    render_keylog();
 }
 
 bool oled_task_user(void) {
-  update_log();
   if (is_keyboard_master()) {
     render_status_main();  // Renders the current keyboard state (layer, lock, caps, scroll, etc)
   } else {
@@ -280,53 +227,4 @@ bool oled_task_user(void) {
   }
     return false;
 }
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-    if (record->event.pressed) {
-        add_keylog(keycode);
-    }
-    return true;
-}
 #endif // OLED_ENABLE
-
-
-// Rotary encoder related code
-#ifdef ENCODER_ENABLE
-bool encoder_update_user(uint8_t index, bool clockwise) {
-  if (index == 0) { // Encoder on master side
-    if(IS_LAYER_ON(_RAISE)) { // on Raise layer
-      // Cursor control
-      if (clockwise) {
-          tap_code(KC_MNXT);
-      } else {
-          tap_code(KC_MPRV);
-      }
-    }
-    else {
-      if (clockwise) {
-          tap_code(KC_VOLU);
-      } else {
-          tap_code(KC_VOLD);
-      }
-    }
-  }
-  else if (index == 1) { // Encoder on slave side
-    if(IS_LAYER_ON(_LOWER)) { // on Lower layer
-      //
-      if (clockwise) {
-          tap_code(KC_RIGHT);
-      } else {
-          tap_code(KC_LEFT);
-      }
-    }
-    else {
-      if (clockwise) {
-          tap_code(KC_DOWN);
-      } else {
-          tap_code(KC_UP);
-      }
-    }
-  }
-    return true;
-}
-#endif
