+++ keyboards/lily58/keymaps/lily58l/rules.mk
@@ -1,1 +1,1 @@
-EXTRAKEY_ENABLE = yes
+TRI_LAYER_ENABLE = yes
