+++ keyboards/lily58/keymaps/mb_via/rules.mk
@@ -1,5 +1,1 @@
 VIA_ENABLE = yes
-OLED_ENABLE = yes
-OLED_DRIVER = SSD1306
-MOUSEKEY_ENABLE = yes
-EXTRAKEY_ENABLE = yes
