+++ keyboards/lily58/keymaps/via/rules.mk
@@ -1,5 +1,2 @@
 VIA_ENABLE = yes
-OLED_ENABLE = yes
-OLED_DRIVER = SSD1306
-MOUSEKEY_ENABLE = yes
-EXTRAKEY_ENABLE = yes
+TRI_LAYER_ENABLE = yes
