+++ keyboards/lily58/light/rules.mk
@@ -1,3 +1,1 @@
-ENCODER_ENABLE = yes       # ENables the use of one or more encoders
-RGBLIGHT_ENABLE = yes      # Enable keyboard RGB light
-LTO_ENABLE = yes 		   # significantly reduce the compiled size, but disable the legacy TMK Macros and Functions features
+# This file intentionally left blank
