+++ keyboards/lily58/r2g/info.json
@@ -1,39 +1,17 @@
 {
     "keyboard_name": "Lily58 R2G",
     "manufacturer": "Mechboards UK",
-    "url": "",
+    "url": "https://mechboards.co.uk/products/lily58-r2g-ready2go-kit",
     "maintainer": "Elliot Powell",
     "processor": "atmega32u4",
     "bootloader": "atmel-dfu",
     "features": {
-        "bootmagic": true,
-        "oled": true,
         "rgb_matrix": true
     },
     "usb": {
         "vid": "0x04D8",
         "pid": "0xEB2E",
         "device_version": "1.0.0"
-    },
-    "diode_direction": "COL2ROW",
-    "matrix_pins": {
-        "rows": ["C6", "D7", "E6", "B4", "B5"],
-        "cols": ["F6", "F7", "B1", "B3", "B2", "B6"]
-    },
-    "split": {
-        "enabled": true,
-        "soft_serial_pin": "D2"
-    },
-    "layouts": {
-        "LAYOUT": {
-            "layout": [
-                {"matrix": [0, 0], "x":0, "y":0.5}, {"matrix": [0, 1], "x":1, "y":0.375}, {"matrix": [0, 2], "x":2, "y":0.125}, {"matrix": [0, 3], "x":3,  "y":0},    {"matrix": [0, 4], "x":4, "y":0.125}, {"matrix": [0, 5], "x":5, "y":0.25},                                                                                               {"matrix": [5, 5], "x":10.5, "y":0.25}, {"matrix": [5, 4], "x":11.5, "y":0.125}, {"matrix": [5, 3], "x":12.5, "y":0}, {"matrix": [5, 2], "x":13.5, "y":0.125}, {"matrix": [5, 1], "x":14.5, "y":0.375}, {"matrix": [5, 0], "x":15.5, "y":0.5},
-                {"matrix": [1, 0], "x":0, "y":1.5}, {"matrix": [1, 1], "x":1, "y":1.375}, {"matrix": [1, 2], "x":2, "y":1.125}, {"matrix": [1, 3], "x":3,  "y":1},    {"matrix": [1, 4], "x":4, "y":1.125}, {"matrix": [1, 5], "x":5, "y":1.25},                                                                                               {"matrix": [6, 5], "x":10.5, "y":1.25}, {"matrix": [6, 4], "x":11.5, "y":1.125}, {"matrix": [6, 3], "x":12.5, "y":1}, {"matrix": [6, 2], "x":13.5, "y":1.125}, {"matrix": [6, 1], "x":14.5, "y":1.375}, {"matrix": [6, 0], "x":15.5, "y":1.5},
-                {"matrix": [2, 0], "x":0, "y":2.5}, {"matrix": [2, 1], "x":1, "y":2.375}, {"matrix": [2, 2], "x":2, "y":2.125}, {"matrix": [2, 3], "x":3,  "y":2},    {"matrix": [2, 4], "x":4, "y":2.125}, {"matrix": [2, 5], "x":5, "y":2.25},                                                                                               {"matrix": [7, 5], "x":10.5, "y":2.25}, {"matrix": [7, 4], "x":11.5, "y":2.125}, {"matrix": [7, 3], "x":12.5, "y":2}, {"matrix": [7, 2], "x":13.5, "y":2.125}, {"matrix": [7, 1], "x":14.5, "y":2.375}, {"matrix": [7, 0], "x":15.5, "y":2.5},
-                {"matrix": [3, 0], "x":0, "y":3.5}, {"matrix": [3, 1], "x":1, "y":3.375}, {"matrix": [3, 2], "x":2, "y":3.125}, {"matrix": [3, 3], "x":3,  "y":3},    {"matrix": [3, 4], "x":4, "y":3.125}, {"matrix": [3, 5], "x":5, "y":3.25}, {"matrix": [4, 5], "x":6, "y":2.75},                   {"matrix": [9, 5], "x":9.5, "y":2.75}, {"matrix": [8, 5], "x":10.5, "y":3.25}, {"matrix": [8, 4], "x":11.5, "y":3.125}, {"matrix": [8, 3], "x":12.5, "y":3}, {"matrix": [8, 2], "x":13.5, "y":3.125}, {"matrix": [8, 1], "x":14.5, "y":3.375}, {"matrix": [8, 0], "x":15.5, "y":3.5},
-                                                                                                                                {"matrix": [4, 1], "x":2.5,"y":4.125},{"matrix": [4, 2], "x":3.5, "y":4.15},{"matrix": [4, 3], "x":4.5,"y":4.25},{"matrix": [4, 4], "x":6, "y":4.25, "h":1.5}, {"matrix": [9, 4], "x":9.5, "y":4.25, "h":1.5}, {"matrix": [9, 3], "x":11, "y":4.25},   {"matrix": [9, 2], "x":12, "y":4.15},    {"matrix": [9, 1], "x":13, "y":4.15}
-            ]
-        }
     },
     "rgb_matrix": {
         "driver": "WS2812",
