+++ keyboards/lily58/r2g/rules.mk
@@ -1,1 +1,1 @@
-LTO_ENABLE = yes
+# This file intentionally left blank
