+++ keyboards/lily58/rev1/info.json
@@ -1,27 +1,25 @@
 {
     "keyboard_name": "Lily58", 
     "manufacturer": "liliums",
-    "url": "", 
+    "url": "https://github.com/kata0510/Lily58", 
     "maintainer": "liliums", 
     "usb": {
         "vid": "0x04D8",
         "pid": "0xEB2D",
         "device_version": "1.0.0"
     },
-    "split": {
-        "soft_serial_pin": "D2"
+    "rgblight": {
+        "led_count": 12,
+        "pin": "D3",
+        "animations": {
+            "rainbow_swirl": true,
+            "static_gradient": true,
+            "twinkle": true
+        }
+    },
+    "features": {
+      "rgblight": true
     },
     "processor": "atmega32u4",
-    "bootloader": "caterina",
-    "layouts": {
-        "LAYOUT": {
-            "layout": [
-                {"x":0, "y":0.5}, {"x":1, "y":0.375}, {"x":2, "y":0.125}, {"x":3, "y":0}, {"x":4, "y":0.125}, {"x":5, "y":0.25}, {"x":10.5, "y":0.25}, {"x":11.5, "y":0.125}, {"x":12.5, "y":0}, {"x":13.5, "y":0.125}, {"x":14.5, "y":0.375}, {"x":15.5, "y":0.5},
-                {"x":0, "y":1.5}, {"x":1, "y":1.375}, {"x":2, "y":1.125}, {"x":3, "y":1}, {"x":4, "y":1.125}, {"x":5, "y":1.25}, {"x":10.5, "y":1.25}, {"x":11.5, "y":1.125}, {"x":12.5, "y":1}, {"x":13.5, "y":1.125}, {"x":14.5, "y":1.375}, {"x":15.5, "y":1.5},
-                {"x":0, "y":2.5}, {"x":1, "y":2.375}, {"x":2, "y":2.125}, {"x":3, "y":2}, {"x":4, "y":2.125}, {"x":5, "y":2.25}, {"x":10.5, "y":2.25}, {"x":11.5, "y":2.125}, {"x":12.5, "y":2}, {"x":13.5, "y":2.125}, {"x":14.5, "y":2.375}, {"x":15.5, "y":2.5},
-                {"x":0, "y":3.5}, {"x":1, "y":3.375}, {"x":2, "y":3.125}, {"x":3, "y":3}, {"x":4, "y":3.125}, {"x":5, "y":3.25}, {"x":6, "y":2.75}, {"x":9.5, "y":2.75}, {"x":10.5, "y":3.25}, {"x":11.5, "y":3.125}, {"x":12.5, "y":3}, {"x":13.5, "y":3.125}, {"x":14.5, "y":3.375}, {"x":15.5, "y":3.5},
-                {"x":2.5, "y":4.125}, {"x":3.5, "y":4.15}, {"x":4.5, "y":4.25}, {"x":6, "y":4.25, "h":1.5}, {"x":9.5, "y":4.25, "h":1.5}, {"x":11, "y":4.25}, {"x":12, "y":4.15}, {"x":13, "y":4.15}
-            ]
-        }
-    }
+    "bootloader": "caterina"
 }
