--- keyboards/lily58/rev1/rev1.c
@@ -1,1 +1,0 @@
-#include "lily58.h"
