+++ keyboards/lily58/rev1/rules.mk
@@ -1,0 +1,1 @@
+# This file intentionally left blank
